requirements.txt
django-user-accounts==1.3.1
pinax-webanalytics==2.0.4
pinax-eventlog==1.1.2
djangorestframework==3.4.3
psycopg2==2.5.4

Django==1.10.2
pinax-theme-bootstrap==7.11.0
